Part of WOR-404: split watcher_finalize.py into finalize.py + helpers with correct patch paths (#832)
